Update README
Add a quickstart script for no-brain execution, but also add a more
verbose explanation for the careful user.

Make directory names more consistent and also keep them in the current
path, not necessarily at $HOME/iyp/.

Check pipeline and commands. Saving a dump to the same file from which
we loaded does not work as that path is not writeable from inside the
container. A separate backups directory is anyways cleaner.