Fix PeeringDB crawlers
PeeringDB 2.46.0 added social media tags for fac, ix, net, and org
objects that may replace the website field entirely at some point in the
future (cf. https://github.com/peeringdb/peeringdb/issues/387).

The social_media field is a list of dicts that can not be flattened by
FlatDict, and using FlatterDict fails if the list is empty. In that
case, a class.FlatterDict object remains in the object, which can not be
pushed to Neo4j.

For this reason we do the flattening ourselves, replacing the
social_media field with social_media_* entries, e.g.,
social_media_website.

Also fixes a circular import caused by requests_cache in ix.py.