Properly recognize that the database has started
Until now we checked the container status, which is always 'running'
once the container is started.
What we really want is to wait on the 'Started.' message output by
Neo4j. The implemented method is not the most elegant, but it works and
prevents us from waiting for two minutes everytime we build the
database.