Add batch_add_properties function and update ip2prefix postprocess
Add a function to add properties to nodes in batches.
This function works only with IDs as I tested it with a node type and
there is no difference in execution time.

The ip2prefix postprocess script now also adds a network and
prefixlen property to each Prefix node, which just contain the network
and prefixlength part separately. This enables easier queries when
looking for most-specific prefixes for example.