Fix BGP.Tools AS names crawler
Due to a format change, the previous version was broken. We now use
pandas to parse the CSV and be robust against future format changes.
The current change also added a class field which is similar to a tag,
which we now also import.
We do not import the country code field, since we do not know where it
comes from and will probably just add redundant information.