Major logging rework
All crawlers now log only using the logging module and not using stderr
anymore. Also the IYP module now automatically logs node/relationship
retrieval and creation. As a consequence, some crawlers are a bit
chatty, which should be fixed in the future.

The general log level was changed to INFO and the level name is now
included in the logs instead of the process name.

As parts of this also some crawlers got reworked:

- alice_lg: Fixed unit test that checked for relationship types which
  are not created.
- ihr:
  - Reference modification time is more precise, matching the exact
    timebin used
  - Data is handled in-memory, removing need for temporary files
  - Moved to batch creation, making the crawler faster and removing
    dependency on other crawlers