Adjust PCH crawler behavior
Until now, we used the first collector in the list to probe the latest
available dataset. However, now happened what I suspected in a code
comment, that only this collector was missing data.

Now we use a beacon collector (at AMS-IX) instead, which has been stable
over the years. Of course, if this gets discontinued at some point, this
will break again.