Add rerun funcionality to postprocess scripts (#195)
Add ability to rerun postprocess scripts

When fixing the dumps we sometimes need to rerun the postprocess scripts
since we added new nodes. For scripts that create relationships, we
first need to delete them before rerunning to avoid duplicates.

This commit adds new delete and rerun functions to postprocess scripts.